refactoring tif processing code and add TifRegionFetcher.py
